+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - removed: `EdgeContext.listUsernames`
 - removed: `EdgeContext.pinLoginEnabled`
 - removed: `listRecoveryQuestionChoices` and related types.
+- removed: `validateMemo` methods and related types.
 
 ## 1.14.0 (2024-01-04)